fix: fixed VBO size declaration + swapped index attribute to integer

+++ src/main/resources/shaders/interface/font/simple.vert
@@ -1,6 +1,6 @@
 #version 430
 
-layout (location = 0) in float vertexId;
+layout (location = 0) in int vertexId;
 layout (location = 1) in vec2 uvPos;
 layout (location = 2) in vec2 uvSize;
 
@@ -13,7 +13,7 @@ out vec2 pass_uv_pos;
 out vec2 pass_uv_size;
 
 void main() {
-    pass_transform = mat4(matrices[int(vertexId)]);
+    pass_transform = mat4(matrices[vertexId]);
     pass_uv_pos = uvPos;
     pass_uv_size = uvSize;
     gl_Position = vec4(0, 0, 0, 1);
